+++ libs/components/src/index.ts
@@ -1,4 +1,9 @@
 // Form Components
 export * from './lib/button/button.component';
+export * from './lib/input/input.component';
+export * from './lib/textarea/textarea.component';
+
+// Layout Components
+export * from './lib/card/card.component';
 
 // More components will be exported as they are implemented